Added more files for evaluated and corrected file qa_pairs.json
File qa_pairs.json had a mistake and should be evaluated again.
File schwierig.json does not have many answers. It's in order to test, if the Chatbot
gives accurate answers to questions that are more difficult to answer.
File unbekannt.json contains answers that cannot  be answered with the context provided.